+++ py/desiutil/test/test_brick.py
@@ -380,6 +380,8 @@ class TestBrick(unittest.TestCase):
                     self.assertTrue((brick_table[n] == dtiling_table[n]).all())
 
     def test_tan_size(self):
+        """Test helper method for converting to WCS.
+        """
         b = B.Bricks(bricksize=1.)
         sz = b.brick_tan_wcs_size()
         self.assertAlmostEqual(sz, 1.05777777, places=6)
@@ -389,6 +391,7 @@ class TestBrick(unittest.TestCase):
         b = B.Bricks(bricksize=4.)
         sz = b.brick_tan_wcs_size()
         self.assertAlmostEqual(sz, 4.23111111, places=6)
+
 
 def test_suite():
     """Allows testing of only this module with the command::